Moved test data for TestRewards.js next to this file
In the previous commit I moved all reward contracts test to a dedicated
directory but I forgot to move there rewardsData.js. Fixing it now.
